--- betatest/templates/betatest/registration_form_field.html
@@ -1,16 +1,0 @@
-<div class="row">
-    <div class="large-5 large-offset-2 columns end">
-        {{ field.errors }}
-    </div>
-</div>
-<div class="row">
-    <div class="large-2 columns">
-        {{ field.label_tag }}
-    </div>
-    <div id="{{ field.id_for_label }}_container" class="large-5 columns">
-        {{ field }}
-    </div>
-    <div class="large-5 columns">
-        <p class="help-text">{{ field.help_text }}</p>
-    </div>
-</div>
