--- betatest/templates/betatest/success.html
@@ -1,13 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Registration Successful{% endblock title %}
-
-{% block content %}
-<div class="container content">
-    <p>Thank you for registering for the OpenCraft beta! We will contact you
-    soon on the status of your application.</p>
-    <p>You will receive verification emails at both of the email addresses you
-    provided. Please click on the links in the emails to confirm your
-    registration.</p>
-</div>
-{% endblock content %}
