+++ requirements.txt
@@ -15,8 +15,7 @@ debtcollector==1.3.0
 decorator==4.0.9
 dj-static==0.0.6
 Django==1.9.5
-# Can be replaced with django-angular==0.8.3 when released to PyPI.
-git+https://github.com/jrief/django-angular.git@c21d7f013fa5108ba7ecc3c7a33de0251aafc604#egg=django-angular==0.8.3
+git+https://github.com/open-craft/django-angular.git@063aad2e5f3cfbd16d300db8b3fce3dbb053d7ce#egg=django-angular==0.8.3
 django-appconf==1.0.2
 django-compressor==2.0
 django-debug-toolbar==1.4
