+++ templates/header.html
@@ -1,0 +1,17 @@
+<header>
+    <nav class="top-bar" role="navigation">
+        <ul class="title-area">
+            <li class="name"><h1><a href="{% url 'index' %}">OpenCraft</a></h1></li>
+        </ul>
+        <section class="top-bar-section">
+            <ul class="right">
+                {% if user.is_authenticated %}
+                    <li><a href="#">{{ user.email }}</a></li>
+                    <li><a href="{% url 'registration:logout' %}">Logout</a></li>
+                {% else %}
+                    <li><a href="{% url 'registration:login' %}">Login</a></li>
+                {% endif %}
+            </ul>
+        </section>
+    </div>
+</header>
